Implement a new approach to extract samples from files
Set "selectSampleExtractionMethod" in your xml to either
- version_1 to select the old method OR
- version_2 to select the new method

Test cases can be found in COMetaDataAccessorSpec:
=> def "Version_2: Extract sample name from BAM basename"
Feel free to take a deep look into this, so that you exactly know, what
happens how.

version_2 of the algorithm will use the values in the lists with the
possible tumor and control sample names. The lists are first joined,
then reverted and then used to match the given filename. This way, we
get try to match for the most specific entry. Note, that the separator
character for metadata in the filename is "_", though you can safely
use "_" in the sample names. E.g. control_some is a valid sample name!

There are three boolean switches, which you can use to modify the
matching behaviour:
- matchExactSampleNames - which will force the matcher to attach an
  "_" at the end of the sample name, so that e.g.
   control => control_ OR
   tumor => tumor_
  By default, the matcher will use the configured sample names as a
  prefix (the names of the list already say it.) This disables it.
- useAllLowerCaseSampleNames
  The name is bad and will be changed! It instructs the matcher to
  use run lowercase on the filename before the matching starts.
- allowSampleTerminationWithIndex
  This will tell the matcher, that the sample name might be followed by
  an integer field, e.g. like:
    control_02_TESTPID
  if you search with "cont", it will find control_02 anyway.

Added accessors for the three configuration values to COConfig.